+++ lib/features/app/app.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../debug/presentation/debug_screen.dart';
+
 /// The main application widget.
 class FlutterStarterTemplate extends StatelessWidget {
   /// Creates a [FlutterStarterTemplate] widget.
@@ -7,12 +9,6 @@ class FlutterStarterTemplate extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Text('Flutter Starter Template'),
-        ),
-      ),
-    );
+    return const MaterialApp(home: DebugScreen());
   }
 }