+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented base Clean Architecture structure (`core`, `features`, `uikit`) with `data`, `domain`, and `presentation` layers.
 - Added all project dependencies from `pubspec.yaml`.
 - Integrated `GetIt` for dependency injection.
+- Added `DebugScreen` widget for testing and debugging purposes.
 
 ### Changed
 - Added `public_member_api_docs` rule to `analysis_options.yaml`.